Changed types back to ints only because it looks nicer

+++ app/src/main/java/com/conradcrates/curvetechnicalchallenge/CalculationModel.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 public interface CalculationModel {
 
-    double doCalculations(List<Double> values);
+    int doCalculations(List<Integer> values);
 }
